Creates some common filter components/functions
